Subtitles: use worldPositionStays=false when moving UI item
This seems to be considered a best practice for reparenting a
RectTransform. By itself this change shouldn't make any difference
in behavior, but it improves interaction with other plugins that set
the renderMode of the canvas to ScreenSpaceCamera. Without this
change, the presence of such a plugin causes subtitles to have the
(very) wrong scale.