Reduced radius of AI/HS2 colliders
Reduced the radius of all AI/HS2 colliders to 0.05. This assists with visible distortion in larger breasts in H scenes when using animations that have the arms close to the body (eg Standing, Standing Missionary/Lifting, Table/Desk Missionary, and Femdom Cowgirl, all in "low" variant) but does not correct it entirely (particularly past 90 size). Some clipping is introduced. NB This has no effect on AI as the existing colliders are active with this plugin and they have a larger radius.

The issue preventing a more complete solution appears to stem from the H scene skeleton having the arm bones in a different place to everywhere else in the game (and studio). As I lack the tools to visualise them in-scene I can only speculate, but it appears as though they are further forward and towards the centre line than they should be. As the bones rotate for some animations, offsets cannot be used to reliably correct for the disparity.

I have included remarks containing the default AI colliders for posterity. If the skeleton issue can be identified and corrected these would be better values than the ones I have set.