[ME] [KK/KKS] Fix hair edits not loading in Maker (#196)
For some reason, when in Maker, KKAPI will fire OnReload immediately after LoadFile, instead of ReloadAsync. So when ReloadAsync does finally run, it'll run ChangeHairAsync, which is hooked by ChangeHairEvent, clearing all hair material edits.

Instead I've decided to hook functions that only run when a hair is changed using maker UI.

Only tested in KK but I assume it works in EC and KKS. The old code works for HS2 so I didn't change it.