[HCharaAdjustment] Fix man's position not being reset when changing positions (#197)
Fixed that when changing the position, the male character was not the target of the reset.

The reproduction procedure is as follows
1. start the free H mode of the game
2. start any sex
3. press O and I keys to move the character
4. move to another location
5. start any sex
6. the man appears in his previous position

The process at the beginning of Update does not reset the position of the guide object until the guide object is hidden.
When the position is changed by in-game operations, the process to hide the guide object was only running on the heroine. This fix extends that process to all controllers.