[ME] Fix timeline compatibility for characters when the character identifier changes (copy character, import scene) (#210)
ME saves the full object path when referencing character game objects (e.g. CommonSpace/chaF_001/BodyTop). But when copying a character or importing a scene with a character, part of this path can change. This fix now works around that limitation (something I completely missed when I implemented it)

Also fixes null reference errors that happen when the scene is reset, and in some cases when the scene is imported. Is done by adding some more null checks in CheckIntegrity()