Fix texture leaks (#218)
Fixed code that was not releasing generated temporary textures. Please merge.
PoseTool seems to be calling UnloadUnused at the caller. So it is not affected. However, I fixed it because of bad manners.

Since this is a resource that will be recovered by UnloadUnusedAssets, there will be no long-term impact from the modification. In the short term, memory usage peaks should be somewhat reduced when textures are changed at high frequency.