Added texture manager - Share identical textures across characters (#229)
Advantages:
Load times are reduced by several tens of percent in the case of scenes with many identical characters. Memory usage is also reduced by several hundred MB.
In the scene I tested, the load time was reduced from 75 seconds to 50 seconds.

Disadvantages:
If there are plug-ins that change textures individually, it will not work correctly.
If there are no identical textures, there is no change. In fact, it slows things down slightly.
When there is a code that forgets to Dispose the TextureContainer, the texture will not disappear with UnloadUnusedAssets.